Detection of folder hierarchy changes in Ping
ActiveSync clients are expected to send FolderSync request ocassionally.
Some send it more some less often. It's hard to predict the interval. Outlook
for example does not send it even on strart.

Ping requests on the other hand are send in specified and predictable intervals.
So, the idea is to detect folder hierarchy changes in Ping request and if that
is a case return with specific error code to cause FolderSync.

Until now we have been detecting only removal of Pinged folders. Now we can detect any
changes, i.e. new folders.

The change provides a bootstrap code that have to be implemented by backend classes.

Change-Id: I624d6b4682804f40a72ec64515487be0f4069668
Reviewed-on: https://gerrit.tine20.org/syncroton/309
Tested-by: jenkins
Reviewed-by: Philipp Schüle <p.schuele@metaways.de>