Add option to limit max number of folders in Sync/Ping request
When using Outlook we observed that it can send Sync requests
for all existing folders, which could be a few hundred.
In such a case the Sync takes a few minutes, because we have to
check all folders, even if they already have been synced. This is
slowing down the whole process significantly.

Change-Id: Id020b3dd50bbe68f0b70482df38bd72f0670ec57
Reviewed-on: https://gerrit.tine20.org/syncroton/313
Tested-by: Philipp Schüle <p.schuele@metaways.de>
Reviewed-by: Philipp Schüle <p.schuele@metaways.de>